Fix integer type inconsistencies for size values
- Changed sgx_quote_buffer_t.length from int to size_t
- Changed verify_sgx_quote quote_len parameter from int to size_t
- Removed unused variables (data, data_size) from echeck_extract_quote
- Updated format specifiers to use %zu for size_t values

Using consistent unsigned types for sizes prevents potential integer
overflow or truncation issues.

+++ src/include/sgx_quote_verify.h
@@ -11,7 +11,7 @@
 typedef echeck_verification_result_t sgx_verification_result_t;
 
 /* Verify SGX quote using built-in CA certificates */
-int verify_sgx_quote(const unsigned char *quote_data, int quote_len,
+int verify_sgx_quote(const unsigned char *quote_data, size_t quote_len,
                      echeck_verification_result_t *result);
 
 /* Verify quote signature using public key from certificate */
